+++ src/components/SlideUpTransition.tsx
@@ -1,0 +1,10 @@
+import { Slide } from "@mui/material";
+import { TransitionProps } from "@mui/material/transitions";
+import { forwardRef } from "react";
+
+export default forwardRef(function Transition(
+  props: TransitionProps & { children: React.ReactElement },
+  ref: React.Ref<unknown>
+) {
+  return <Slide ref={ref} direction="up" {...props} />;
+});
