+++ src/features/portfolio/components/PortfolioList/desktop/PortfolioListTableToolBar.tsx
@@ -7,7 +7,7 @@ import { useBoolean } from "@fineants/demolition";
 import { Toolbar, Tooltip, Typography } from "@mui/material";
 import designSystem from "@styles/designSystem";
 import styled from "styled-components";
-import PortfolioAddDialog from "../../PortfolioAddOrEditDialog/PortfolioAddOrEditDialog";
+import PortfolioAddDialog from "../../PortfolioAddOrEditDialog/desktop/PortfolioAddOrEditDialogD";
 
 interface PortfolioListTableToolBarProps {
   selected: readonly PortfolioItem[];
